# doc/swagger.yaml
definitions:
...
  credential.VerifiablePresentation:
    properties:
      '@context':
        description: Either a string or set of strings
      holder:
        type: string
      id:
        type: string
      presentation_submission:
        description: an optional field as a part of https://identity.foundation/presentation-exchange/#embed-targets
      proof: {}
      type: {}
      verifiableCredential:
        description: Verifiable credential could be our object model, a JWT, or any
          other valid credential representation
        items: {}
        type: array
    required:
    - type
    type: object
  crypto.KeyType:
    enum:
    - Ed25519
    - X25519
    - secp256k1
    - secp256k1-ECDSA
    - P-224
    - P-256
    - P-384
    - P-521
    - RSA
    - BLS12381G1
    - BLS12381G2
    - Dilithium2
    - Dilithium3
    - Dilithium5
    type: string
    x-enum-varnames:
    - Ed25519
    - X25519
    - SECP256k1
    - SECP256k1ECDSA
    - P224
    - P256
    - P384
    - P521
    - RSA
    - BLS12381G1
    - BLS12381G2
    - Dilithium2
    - Dilithium3
    - Dilithium5
...
  jwx.PublicKeyJWK:
    properties:
      alg:
        type: string
      crv:
        type: string
      e:
        type: string
      key_ops:
        type: string
      kid:
        type: string
      kty:
        type: string
      "n":
        type: string
      use:
        type: string
      x:
        type: string
      "y":
        type: string
    required:
    - kty
    type: object
...
  pkg_server_router.GetKeyDetailsResponse:
    properties:
      controller:
        type: string
      createdAt:
        description: Represents the time at which the key was created. Encoded according
          to RFC3339.
        type: string
      id:
        type: string
      publicKeyJwk:
        allOf:
        - $ref: '#/definitions/jwx.PublicKeyJWK'
        description: |-
          The public key in JWK format according to RFC7517. This public key is associated with the private
          key with the associated ID.
      type:
        $ref: '#/definitions/crypto.KeyType'
    type: object
...
  pkg_server_router.GetSubmissionResponse:
    properties:
      reason:
        description: The reason why the submission was approved or denied.
        type: string
      status:
        description: One of {`pending`, `approved`, `denied`, `cancelled`}.
        type: string
      verifiablePresentation:
        allOf:
        - $ref: '#/definitions/credential.VerifiablePresentation'
        description: The verifiable presentation containing the presentation_submission
          along with the credentials presented.
    required:
    - status
    type: object
...
  pkg_server_router.ListOperationsResponse:
    properties:
      nextPageToken:
        description: Pagination token to retrieve the next page of results. If the
          value is "", it means no further results for the request.
        type: string
      operations:
        items:
          $ref: '#/definitions/pkg_server_router.Operation'
        type: array
    type: object
...
  pkg_server_router.Operation:
    properties:
      done:
        description: Whether this operation has finished.
        type: boolean
      id:
        description: The name of the resource related to this operation. E.g. "presentations/submissions/<uuid>"
        type: string
      result:
        allOf:
        - $ref: '#/definitions/pkg_server_router.OperationResult'
        description: Populated if Done == true.
    required:
    - done
    - id
    type: object
  pkg_server_router.OperationResult:
    properties:
      error:
        description: Populated when there was an error with the operation.
        type: string
      response:
        description: Populated iff Error == "". The type should be specified in the
          calling APIs documentation.
    type: object
...
  pkg_server_router.RevokeKeyResponse:
    properties:
      id:
        type: string
    type: object
...
  pkg_server_router.StoreKeyRequest:
    properties:
      base58PrivateKey:
        description: Base58 encoding of the bytes that result from marshalling the
          private key using golang's implementation.
        type: string
      controller:
        description: See https://www.w3.org/TR/did-core/#did-controller
        type: string
      id:
        description: |-
          The `id` field is the unique identifier for this object. If set to a resolvable DID, the ssi-service will use
          the private key encoded in the `PrivateKeyBase58` field of this object to sign objects issued or authored by this
          DID; otherwise, it will only be used to identify this object.
        type: string
      type:
        allOf:
        - $ref: '#/definitions/crypto.KeyType'
        description: |-
          Identifies the cryptographic algorithm family used with the key.
          One of the following: "Ed25519", "X25519", "secp256k1", "P-224", "P-256", "P-384", "P-521", "RSA".
    required:
    - base58PrivateKey
    - controller
    - id
    - type
    type: object
...
info:
  contact:
    email: tbd-developer@squareup.com
    name: CYWARE
    url: https://github.com/cyware/ssi-service/issues
  description: 'The Self Sovereign Identity Service: Managing DIDs, Verifiable Credentials,
    and more!'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: 0.0.3
paths:
...
  /v1/keys:
    put:
      consumes:
      - application/json
      description: Stores a key to be used by the service
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.StoreKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Store a keys
      tags:
      - KeyStore
  /v1/keys/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        Marks a key as being revoked, along with the timestamps of when it was revoked.
        NB: the key can still be used for signing. This will likely be addressed before v1 is released.
      parameters:
      - description: ID of the key to revoke
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.RevokeKeyResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Revoke a key
      tags:
      - KeyStore
    get:
      consumes:
      - application/json
      description: Get details about a stored key
      parameters:
      - description: ID of the key to get
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.GetKeyDetailsResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get details for a key
      tags:
      - KeyStore
...
  /v1/operations:
    get:
      consumes:
      - application/json
      description: List operations according to the request
      parameters:
      - description: 'The name of the parent''s resource. For example: `?parent=/presentation/submissions`'
        in: query
        name: parent
        type: string
      - description: 'A standard filter expression conforming to https://google.aip.dev/160.
          For example: `?filter=done='
        in: query
        name: filter
        type: string
      - description: Hint to the server of the maximum elements to return. More may
          be returned. When not set, the server will return all elements.
        in: query
        name: pageSize
        type: number
      - description: Used to indicate to the server to return a specific page of the
          list results. Must match a previous requests' `nextPageToken`.
        in: query
        name: pageToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.ListOperationsResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List operations
      tags:
      - Operations
  /v1/operations/{id}:
    get:
      consumes:
      - application/json
      description: Get operation by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.Operation'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get an operation
      tags:
      - Operations
  /v1/operations/cancel/{id}:
    get:
      consumes:
      - application/json
      description: Cancels an active operation, if possible.
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.Operation'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Cancel an operation
      tags:
      - Operations
...
  /v1/presentations/submissions/{id}:
    get:
      consumes:
      - application/json
      description: Get a Presentation Submission by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.GetSubmissionResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get a Presentation Submission
      tags:
      - PresentationSubmissions
...
swagger: "2.0"